Split character spell routes
