+++ src/domains/characters/runtime/routes.ts
@@ -15,20 +15,13 @@ import {
 import {
 	CreateCharacterRequestSchema,
 	RestoreCharacterSpellSlotRequestSchema,
-	SaveCharacterSpellRequestSchema,
-	SearchCharacterSpellsRequestSchema,
 	UpdateCharacterHealthRequestSchema,
 	UpdateCharacterLevelRequestSchema,
 	UpdateCharacterSpellSlotsRequestSchema,
 	UseCharacterSpellSlotRequestSchema,
 } from "../types/index.js";
-import {
-	parseBody,
-	parseParams,
-	parseSpellParams,
-	sendSpellError,
-	sendSpellSlotError,
-} from "./route-helpers.js";
+import { parseBody, parseParams, sendSpellSlotError } from "./route-helpers.js";
+import { registerCharacterSpellRoutes } from "./routes.spells.js";
 
 const defaultCharacterService = createCharacterService();
 const defaultCharacterHealthService = createCharacterHealthService();
@@ -220,88 +213,8 @@ export async function registerCharacterRoutes(
 		}
 	});
 
-	app.get("/api/characters/:characterId/spells", async (request, reply) => {
-		const params = parseParams(request, reply);
-		if (!params) return;
-
-		const currentUser = await getCurrentUser(request, reply);
-		try {
-			return await characterSpellService.listCharacterSpells(
-				currentUser.user.id,
-				params.characterId,
-			);
-		} catch (error) {
-			return sendSpellError(error, reply);
-		}
-	});
-
-	app.get("/api/characters/:characterId/spells/:spellId", async (request, reply) => {
-		const params = parseSpellParams(request, reply);
-		if (!params) return;
-
-		const currentUser = await getCurrentUser(request, reply);
-		try {
-			return await characterSpellService.getCharacterSpellDetails(
-				currentUser.user.id,
-				params.characterId,
-				params.spellId,
-			);
-		} catch (error) {
-			return sendSpellError(error, reply);
-		}
-	});
-
-	app.post("/api/characters/:characterId/spells/search", async (request, reply) => {
-		const params = parseParams(request, reply);
-		if (!params) return;
-
-		const body = parseBody(SearchCharacterSpellsRequestSchema, request.body, reply);
-		if (!body) return;
-
-		const currentUser = await getCurrentUser(request, reply);
-		try {
-			return await characterSpellService.searchCharacterSpells(
-				currentUser.user.id,
-				params.characterId,
-				body,
-			);
-		} catch (error) {
-			return sendSpellError(error, reply);
-		}
-	});
-
-	app.post("/api/characters/:characterId/spells", async (request, reply) => {
-		const params = parseParams(request, reply);
-		if (!params) return;
-
-		const body = parseBody(SaveCharacterSpellRequestSchema, request.body, reply);
-		if (!body) return;
-
-		const currentUser = await getCurrentUser(request, reply);
-		try {
-			return await characterSpellService.saveCharacterSpell(
-				currentUser.user.id,
-				params.characterId,
-				body,
-			);
-		} catch (error) {
-			return sendSpellError(error, reply);
-		}
-	});
-
-	app.delete("/api/characters/:characterId/spells/:spellId", async (request, reply) => {
-		const params = parseSpellParams(request, reply);
-		if (!params) return;
-
-		const currentUser = await getCurrentUser(request, reply);
-		try {
-			return await characterSpellService.removeCharacterSpell(
-				currentUser.user.id,
-				params.characterId,
-				params.spellId,
-			);
-		} catch (error) {
-			return sendSpellError(error, reply);
-		}
+	await registerCharacterSpellRoutes(app, {
+		characterSpellService,
+		getCurrentUser,
 	});
 }
